implemented history and request queue
history is specific to each request id, which is an option that can be
sent over the request. By default the helper function sets the id to the
script name, but the main script uses 'mpv' as the default id. The
ids are shared between scripts, which means similar scripts can pool
their histories.

The queue system allows multiple user requests to be sent at once.
Requests will be held in the queue until the user enters or cancels a
request. By default multiple requests with the same id cannot be queued,
however this can be enabled using the queuable option flag.